requirements.txt
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
Pillow==4.3.0
pytesseract==0.1.7
